added schema; moved imports around
